requirements.txt
# AI Travel Assistant + Business Analytics Dashboard
# Python 3.11+ (recommended: 3.12 or 3.13)

# --- UI ---
streamlit>=1.55.0

# --- AI ---
openai>=1.30.0
google-generativeai>=0.5.4

# --- Database ---
pymongo>=4.7.2
dnspython>=2.6.1         # needed for MongoDB Atlas SRV URIs

# --- Auth ---
bcrypt>=4.1.3

# --- Data & Charts ---
plotly>=5.22.0
pandas>=3.0.2
numpy>=2.0.0
faker>=25.2.0            # realistic mock data generation

# --- Utilities ---
python-dotenv>=1.0.1
cachetools>=5.3.3

# --- Optional but recommended ---
streamlit-chat>=0.1.1    # chat message components